requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0
scikit-learn>=1.1.0
opencv-python>=4.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.64.0
pandas>=1.5.0
streamlit>=1.28.0